+++ packages/typegpu/tests/examples/individual/wgsl-resolution.test.ts
@@ -90,7 +90,7 @@ describe('wgsl resolution example', () => {
         var cohesion = vec2f();
         var alignmentCount = 0;
         var cohesionCount = 0;
-        for (var i = u32(0); (i < arrayLength(&currentTrianglePos_10)); i++) {
+        for (var i = 0u; (i < arrayLength(&currentTrianglePos_10)); i++) {
           if ((i == index)) {
             continue;
           }
